// angular-netflix/src/app/movie.interface.ts
export interface Movie {
    id: number;
    title: string;
    overview: string;
    image: string;
  }
  
  export interface MoiveDetail {
    id: number;
    title: string;
    overview: string;
    rating: number;
    releaseDate: string;
    homepage: string;
    genres: string[];
    image: string;
  }
  